+++ pkg/sql/tests/BUILD.bazel
@@ -41,6 +41,7 @@ go_test(
         "split_test.go",
         "system_table_test.go",
         "table_split_test.go",
+        "tracing_sql_test.go",
         "virtual_table_test.go",
     ],
     data = glob(["testdata/**"]),
@@ -89,6 +90,7 @@ go_test(
         "//pkg/util/randutil",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
+        "//pkg/util/tracing",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_cockroach_go//crdb",
         "@com_github_cockroachdb_datadriven//:datadriven",
